requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   django
    #   django-htmx
asttokens==2.4.1
    # via stack-data
async-timeout==4.0.3
    # via redis
bcrypt==4.1.2
    # via paramiko
beautifulsoup4==4.12.3
    # via django-check-seo
black==24.2.0
    # via -r requirements.in
boto3==1.34.49
    # via -r requirements.in
botocore==1.34.49
    # via
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
cfgv==3.4.0
    # via pre-commit
chardet==5.2.0
    # via reportlab
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   django-tasks-scheduler
    #   rq
cmsplugin-filer @ git+https://github.com/cruncher/cmsplugin-filer@0c3755cad42fa713cc75017d0322270c888a90a3
    # via -r requirements.in
croniter==2.0.1
    # via django-tasks-scheduler
cryptography==42.0.5
    # via
    #   fido2
    #   paramiko
cssselect2==0.7.0
    # via svglib
decorator==5.1.1
    # via ipython
distlib==0.3.8
    # via virtualenv
django==4.2.10
    # via
    #   -r requirements.in
    #   cmsplugin-filer
    #   django-appconf
    #   django-classy-tags
    #   django-cms
    #   django-extensions
    #   django-filer
    #   django-formtools
    #   django-htmx
    #   django-otp
    #   django-parler
    #   django-passkeys
    #   django-polymorphic
    #   django-rosetta
    #   django-sekizai
    #   django-storages
    #   django-tasks-scheduler
    #   django-treebeard
    #   djangocms-admin-style
    #   djangocms-versioning
    #   easy-thumbnails
django-admin-honeypot @ git+https://github.com/cruncher/django-admin-honeypot.git@1cf94e3d9809bb45e54fa83f9587a8c9bcc62fdf
    # via -r requirements.in
django-appconf==1.0.6
    # via cmsplugin-filer
django-check-seo==1.0.0
    # via -r requirements.in
django-classy-tags==4.1.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==4.1.0
    # via
    #   -r requirements.in
    #   cmsplugin-filer
    #   djangocms-attributes-field
    #   djangocms-page-meta
    #   djangocms-snippet
    #   djangocms-text-ckeditor
    #   djangocms-versioning
django-eml-email-backend==0.1
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filer==3.1.1
    # via
    #   -r requirements.in
    #   cmsplugin-filer
    #   djangocms-page-meta
django-formtools==2.5.1
    # via django-cms
django-fsm==2.8.1
    # via djangocms-versioning
django-htmx==1.17.2
    # via -r requirements.in
django-impersonate==1.9.2
    # via -r requirements.in
django-ipware==6.0.4
    # via django-admin-honeypot
django-meta==2.4.2
    # via djangocms-page-meta
django-otp==1.3.0
    # via -r requirements.in
django-parler==2.3
    # via -r requirements.in
django-passkeys @ git+https://github.com/halitcelik/django-passkeys.git@2f8439e9435a12bda3ca6779dffa22be4506089e
    # via -r requirements.in
django-polymorphic==3.1.0
    # via django-filer
django-querycount==0.8.3
    # via -r requirements.in
django-rosetta==0.10.0
    # via -r requirements.in
django-sekizai==4.1.0
    # via
    #   -r requirements.in
    #   cmsplugin-filer
    #   django-cms
django-storages==1.14.2
    # via -r requirements.in
django-tasks-scheduler==1.3.1
    # via -r requirements.in
django-treebeard==4.7.1
    # via django-cms
djangocms-admin-style==3.3.0
    # via django-cms
djangocms-attributes-field==3.0.0
    # via cmsplugin-filer
djangocms-page-meta==1.4.3
    # via django-check-seo
djangocms-snippet==3.1.1
    # via -r requirements.in
djangocms-text-ckeditor==5.1.5
    # via -r requirements.in
djangocms-versioning==2.0.0
    # via -r requirements.in
easy-thumbnails[svg]==2.8.5
    # via
    #   -r requirements.in
    #   cmsplugin-filer
    #   django-filer
executing==2.0.1
    # via stack-data
fabric3 @ git+https://github.com/cruncher/fabric.git@1c60d21b37af4cdf2ce9cc6ef334cc7a978ee3e9
    # via -r requirements.in
fido2==1.1.1
    # via django-passkeys
filelock==3.13.1
    # via virtualenv
gunicorn==21.2.0
    # via -r requirements.in
html5lib==1.1
    # via djangocms-text-ckeditor
identify==2.5.35
    # via pre-commit
idna==3.6
    # via requests
ipython==8.22.1
    # via -r requirements.in
jedi==0.19.1
    # via
    #   ipython
    #   pudb
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==5.1.0
    # via
    #   django-check-seo
    #   svglib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via werkzeug
matplotlib-inline==0.1.6
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
packaging==23.2
    # via
    #   black
    #   django-cms
    #   djangocms-text-ckeditor
    #   gunicorn
    #   pudb
paramiko==2.12.0
    # via fabric3
parso==0.8.3
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pillow==10.2.0
    # via
    #   -r requirements.in
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
platformdirs==4.2.0
    # via
    #   black
    #   virtualenv
polib==1.2.0
    # via django-rosetta
pre-commit==3.6.2
    # via -r requirements.in
prompt-toolkit==3.0.43
    # via ipython
psycopg2-binary==2.9.9
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pudb==2024.1
    # via -r requirements.in
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.17.2
    # via
    #   ipython
    #   pudb
    #   rich
pymemcache==4.0.0
    # via -r requirements.in
pynacl==1.5.0
    # via paramiko
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   botocore
    #   croniter
python-ipware==2.0.1
    # via django-ipware
python-memcached==1.62
    # via -r requirements.in
pytz==2024.1
    # via
    #   -r requirements.in
    #   croniter
pyyaml==6.0.1
    # via pre-commit
qrcode==7.4.2
    # via -r requirements.in
raven==6.10.0
    # via -r requirements.in
redis==5.0.1
    # via
    #   -r requirements.in
    #   rq
reportlab==4.1.0
    # via
    #   easy-thumbnails
    #   svglib
requests==2.31.0
    # via
    #   -r requirements.in
    #   django-check-seo
    #   django-rosetta
rich==13.7.0
    # via -r requirements.in
rq==1.16.0
    # via django-tasks-scheduler
s3transfer==0.10.0
    # via boto3
sentry-sdk==1.40.5
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   fabric3
    #   html5lib
    #   paramiko
    #   python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stack-data==0.6.3
    # via ipython
svglib==1.5.1
    # via easy-thumbnails
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.10.0
    # via
    #   qrcode
    #   urwid
ua-parser==0.18.0
    # via
    #   django-passkeys
    #   user-agents
unidecode==1.3.8
    # via
    #   -r requirements.in
    #   django-check-seo
urllib3==2.0.7
    # via
    #   botocore
    #   requests
    #   sentry-sdk
urwid==2.6.4
    # via
    #   pudb
    #   urwid-readline
urwid-readline==0.14
    # via pudb
user-agents==2.2.0
    # via django-passkeys
virtualenv==20.25.1
    # via pre-commit
wcwidth==0.2.13
    # via
    #   prompt-toolkit
    #   urwid
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
werkzeug==3.0.1
    # via -r requirements.in
xlrd==2.0.1
    # via -r requirements.in
xlwt==1.3.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
